+++ src/main/java/oripa/util/Pair.java
@@ -1,0 +1,54 @@
+/**
+ * ORIPA - Origami Pattern Editor
+ * Copyright (C) 2013-     ORIPA OSS Project  https://github.com/oripa/oripa
+ * Copyright (C) 2005-2009 Jun Mitani         http://mitani.cs.tsukuba.ac.jp/
+
+    This program is free software: you can redistribute it and/or modify
+    it under the terms of the GNU General Public License as published by
+    the Free Software Foundation, either version 3 of the License, or
+    (at your option) any later version.
+
+    This program is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+    GNU General Public License for more details.
+
+    You should have received a copy of the GNU General Public License
+    along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+package oripa.util;
+
+import java.util.Objects;
+
+public class Pair<V1, V2> {
+	private final V1 v1;
+	private final V2 v2;
+
+	public Pair(final V1 v1, final V2 v2) {
+		this.v1 = v1;
+		this.v2 = v2;
+	}
+
+	public V1 getV1() {
+		return v1;
+	}
+
+	public V2 getV2() {
+		return v2;
+	}
+
+	@Override
+	public boolean equals(final Object obj) {
+		if (!(obj instanceof Pair)) {
+			return false;
+		}
+		@SuppressWarnings("unchecked")
+		var o = (Pair<V1, V2>) obj;
+		return v1.equals(o.v1) && v2.equals(o.v2);
+	}
+
+	@Override
+	public int hashCode() {
+		return Objects.hash(v1, v2);
+	}
+}
